Update Go to latest version
The action is able to read the go.mod version to discover what go
version to install. So lets use that just like in ironlib to avoid
needing to keep multiple files in sync.

Needed to update golangci-lint because the previous version fails. I
used the same version as in ironlib. New golangci-lint complained about
some new stuff, so I fixed it.